# [POC] PERF: 1d version of the cython grouped aggregation algorithms