--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -702,6 +702,59 @@ group_mean_float64 = _group_mean['double']
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
+def _group_mean_1d(floating[:] out,
+                   int64_t[:] counts,
+                   ndarray[floating, ndim=1] values,
+                   const int64_t[:] labels,
+                   Py_ssize_t min_count=-1):
+    cdef:
+        Py_ssize_t i, j, N, K, lab, ncounts = len(counts)
+        floating val, count, y, t
+        floating[:] sumx, compensation
+        int64_t[:] nobs
+        Py_ssize_t len_values = len(values), len_labels = len(labels)
+
+    assert min_count == -1, "'min_count' only used in add and prod"
+
+    if len_values != len_labels:
+        raise ValueError("len(index) != len(labels)")
+
+    nobs = np.zeros((<object>out).shape, dtype=np.int64)
+    sumx = np.zeros_like(out)
+    compensation = np.zeros_like(out)
+
+    N, = (<object>values).shape
+
+    with nogil:
+        for i in range(N):
+            lab = labels[i]
+            if lab < 0:
+                continue
+
+            counts[lab] += 1
+            val = values[i]
+            # not nan
+            if val == val:
+                nobs[lab] += 1
+                y = val - compensation[lab]
+                t = sumx[lab] + y
+                compensation[lab] = t - sumx[lab] - y
+                sumx[lab] = t
+
+        for i in range(ncounts):
+                count = nobs[i]
+                if nobs[i] == 0:
+                    out[i] = NAN
+                else:
+                    out[i] = sumx[i] / count
+
+
+group_mean_1d_float32 = _group_mean_1d['float']
+group_mean_1d_float64 = _group_mean_1d['double']
+
+
+@cython.wraparound(False)
+@cython.boundscheck(False)
 def _group_ohlc(floating[:, :] out,
                 int64_t[:] counts,
                 ndarray[floating, ndim=2] values,