@jreback 2021-02-17
this is a really strange result 
you eliminated a loop over the columns 

i wouldn't expect any difference here

my guess is that the 2d result array memory layout is unfavorable compared to the 1d