@jorisvandenbossche 2021-02-17
The inner part of the function for the 2D version is:

```
                val = values[i, j]
                # not nan
                if val == val:
                    nobs[lab, j] += 1
                    y = val - compensation[lab, j]
                    t = sumx[lab, j] + y
                    compensation[lab, j] = t - sumx[lab, j] - y
                    sumx[lab, j] = t
```

while for the 1D version it is:

```
            val = values[i]
            # not nan
            if val == val:
                nobs[lab] += 1
                y = val - compensation[lab]
                t = sumx[lab] + y
                compensation[lab] = t - sumx[lab] - y
                sumx[lab] = t
```

So all `arr[i, j]` 2D indexing steps are replaced with the simpler `arr[i]`. So I personally don't find it surprising that this gives *some* improvement (how much improvement to expect from it, I have no idea apart from the benchmarking I did, as nothing else changed).


> my guess is that the 2d result array memory layout is unfavorable compared to the 1d

That was my first guess as well (as for example for the plain (non grouped) reductions, this matters), but from testing it with the example above, that doesn't seem to matter in this case:. Explicitly testing the different memory layouts with the 2D algorithm:

```
In [48]: values = df[['a']]._mgr.blocks[0].values.T

In [49]: valuesC = values.copy(order="C")

In [50]: valuesF = values.copy(order="F")

In [51]: %%timeit
    ...: result = np.empty(out_shape, out_dtype)
    ...: result.fill(np.nan)
    ...: counts = np.zeros(len(uniques), dtype=np.int64)
    ...: pd._libs.groupby.group_mean_float64(result, counts, values, codes, min_count=-1)
4.12 ms ± 90.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [52]: %%timeit
    ...: result = np.empty(out_shape, out_dtype)
    ...: result.fill(np.nan)
    ...: counts = np.zeros(len(uniques), dtype=np.int64)
    ...: pd._libs.groupby.group_mean_float64(result, counts, valuesC, codes, min_count=-1)
4.26 ms ± 247 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [53]: %%timeit
    ...: result = np.empty(out_shape, out_dtype)
    ...: result.fill(np.nan)
    ...: counts = np.zeros(len(uniques), dtype=np.int64)
    ...: pd._libs.groupby.group_mean_float64(result, counts, valuesF, codes, min_count=-1)
4.09 ms ± 20 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

And additionally explicitly with how it would work for a 1D column that needs to be reshaped to pass into the algorithm:

```
In [63]: values1d = df['a']._mgr.blocks[0].values.copy()

In [64]: values1d_reshaped = values1d[:, None]

In [65]: %%timeit
    ...: result = np.empty(out_shape, out_dtype)
    ...: result.fill(np.nan)
    ...: counts = np.zeros(len(uniques), dtype=np.int64)
    ...: pd._libs.groupby.group_mean_float64(result, counts, values1d_reshaped, codes, min_count=-1)
4.13 ms ± 55.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

So those all don't really show any significant difference.