@jorisvandenbossche 2021-02-17
And to be explicit: I am testing the 2D (vs 1D) algorithm here for a *single* column (so an array of shape `(N, 1)`). I suppose that the 2D algorithm will still be faster for an actual 2D array from multiple columns, compared to calling the 1D algorithm multiple times on the different columns. 
But what I am benchmarking here is exactly for the case we *would* have our columns stored as 1D arrays, but still need to use the 2D algorithm vs having a 1D algorithm.