@jreback 2021-02-17
i think if u reverse the ordering

eg j, lab this will have same perf (need to transpose when creating memory)

as the indexing is the costly part here