@jorisvandenbossche 2021-02-17
> as the indexing is the costly part here

Exactly, and I reduce the amount of indexing (indexing into 2D vs 1D)

> i think if u reverse the ordering, eg j, lab this will have same perf (need to transpose when creating memory)

Not fully sure what you mean, but I did verify the result for all cases to be identical with the actual `df.groupby().mean()`, so I don't think I did something wrong in the for loop (otherwise I wouldn't get the correct result?) 
And I did transpose the block values in the first case.