@jorisvandenbossche 2021-02-17
> i think if u reverse the ordering, eg j, lab this will have same perf

Ah, I think I understand now. You didn't mean something was wrong in the current example, but that if the 2D algorithm *would* be written to work on data transposed compared to what it is now, it would be faster?

Currently, the values are passed in as array with shape `(n_rows, n_cols)`), and you suggest to try what it gives if they are passed in as `(n_cols, n_rows)` ? (so the indexing like `sumx[lab, j]` would become `sumx[j, lab]`)