@jreback 2021-02-17
yea exactly

try to fill in reversed indexing order to see if that makes the 2d case better