@jreback 2021-02-17
the issue is the cost of the indexing is dependent on the memory layout