@jreback 2021-02-17
and i think the 2d case would be more performing if we allocate 1d contiguous blocks as these will be cache performant

and since then aggregating you don't need to concat then