@jorisvandenbossche 2021-02-17
Will try that. Maybe that it indeed has significant enough impact for the 1D-as-2D array case.